pyproject.toml
[project]
name = "agentic-yaitsiu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "ollama>=0.4.7",
    "openai>=1.60.2",
    "pip>=25.0",
    "polars>=1.21.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "result>=0.17.0",
    "tweepy>=4.15.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
indent-width = 4

[tool.ruff.format]
indent-style = "tab"


